Replaced SearchBar for UserBar
