+++ src/components/UserBar.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import UserProfileButton from './UserProfileButton';
+import SearchBox from './SearchBox';
+
+const UserBar = ({ filter, setFilter, onKeyPress }) => (
+	<div className='pt-2 relative flex items-center text-white'>
+		<UserProfileButton />
+		<SearchBox
+			filter={filter}
+			setFilter={setFilter}
+			onKeyPress={onKeyPress}
+		/>
+	</div>
+);
+
+UserBar.propTypes = {
+	filter: PropTypes.string.isRequired,
+	setFilter: PropTypes.func.isRequired,
+	onKeyPress: PropTypes.func.isRequired,
+};
+
+export default UserBar;
